refactor: 시뮬레이터 영역을 Competition에서 분리
