+++ frontend/src/components/Simulation/index.ts
@@ -1,0 +1,1 @@
+export * from './Simulator';
